Replace largest_angle hack with proper 2d->3d error function
Compute if a line (triangle edge or contour line) needs to be split in
two in a new way. Instead of using a hack estimate with angles, the code
now instead the actual distance from the earth's surface that the line
shortcuts through the sphere
